ml/requirements.txt
torch
torchvision
transformers
Pillow
requests
numpy
onnx
onnxruntime
open_clip_torch
